requirements.txt
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
clickclick==1.2.2
colorama==0.4.0
connexion==1.5.3
Flask==1.0.2
Flask-JWT-Extended==3.13.1
gunicorn==19.9.0
idna==2.7
inflection==0.3.1
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.3.0
pluggy==0.8.0
psycopg2==2.7.5
py==1.7.0
PyJWT==1.6.4
pylint==2.1.1
pytest==3.8.2
PyYAML==3.13
requests==2.20.0
six==1.11.0
swagger-spec-validator==2.4.1
urllib3==1.24
Werkzeug==0.14.1
wrapt==1.10.11
